:recycle: #99 他のComponentと同じように言語情報を受け取り内部で翻訳を行うように改修
